requirements.txt
pyusb>=1.2
opencv-python=4.9.0.80
mediapipe>=0.10